+++ util/output.h
@@ -1,7 +1,7 @@
 #ifndef __OUTPUT__
 #define __OUTPUT__
 
-void sprintHex(unsigned char *vals, size_t sss, char *pResult, unsigned int size);
+void sprintHex(unsigned char *vals, unsigned int sss, char *pResult, unsigned int size);
 
 int printBytes(unsigned char *p, unsigned int size, char *pResult, unsigned int resultLength);
